Editing TableRow component to include alert before clearing table

+++ front-end/src/dashboard/TableRow.js
@@ -1,30 +1,26 @@
 import React from 'react';
 import axios from 'axios';
 
-function TableRow({ table, setTablesReservationChange, tablesReservationChange }) {
+function TableRow({ table, tables, setTables, setTablesError }) {
 
     const { table_id, table_name, capacity, table_availability, assigned_reservation_id } = table;
 
     async function handleFinish() {
         try {
-            await axios.delete(`http://localhost:5000/tables/${table.table_id}/seat`);
-            await axios.put(`http://localhost:5000/reservations/${assigned_reservation_id}/status`, { data: { 'status': 'finished', 'reservation_id': assigned_reservation_id } })
-            // const tablesArray = tables.filter(table => table.table_id !== table_id);
-            // let updatedTable = { ...table, table_availability: 'free' };
-            // let updatedTableArray = [...tablesArray, updatedTable];
-            // updatedTableArray.sort((a, b) => {
-            //     if (a.table_name > b.table_name) {
-            //         return 1;
-            //     } else if (a.table_name < b.table_name) {
-            //         return -1;
-            //     } else {
-            //         return 0;
-            //     };
-            // })
-            setTablesReservationChange(tablesReservationChange + 1);
+            const response = window.confirm("Is this table ready to seat new guests? This cannot be undone.")
+            if (response) {
+                await axios.delete(`http://localhost:5000/tables/${table.table_id}/seat`);
+
+                await axios.put(`http://localhost:5000/reservations/${assigned_reservation_id}/status`, { data: { 'status': 'finished', 'reservation_id': assigned_reservation_id } })
+
+                const tablesArray = tables.filter(table => table.table_id !== table_id);
+                let updatedTable = { ...table, table_availability: 'free' };
+                let updatedTablesArray = [...tablesArray, updatedTable];
+                setTables(updatedTablesArray)
+            }
         }
         catch (error) {
-            console.log(error)
+            setTablesError(error)
         }
     }
